+++ src-tauri/src/sql/stocks.rs
@@ -4,7 +4,7 @@ pub fn create_levels_table() -> &'static str {
     "CREATE TABLE IF NOT EXISTS stock_levels (
         variant_id INTEGER NOT NULL,
         warehouse_id INTEGER NOT NULL,
-        quantity REAL NOT NULL DEFAULT 0,
+        quantity INTEGER NOT NULL DEFAULT 0,
         PRIMARY KEY (variant_id, warehouse_id),
         FOREIGN KEY(variant_id) REFERENCES product_variants(id),
         FOREIGN KEY(warehouse_id) REFERENCES warehouses(id)
@@ -17,7 +17,7 @@ pub fn create_movements_table() -> &'static str {
         variant_id INTEGER NOT NULL,
         from_warehouse_id INTEGER,
         to_warehouse_id INTEGER,
-        quantity REAL NOT NULL,
+        quantity INTEGER NOT NULL,
         \"type\" TEXT NOT NULL,
         created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
         FOREIGN KEY(variant_id) REFERENCES product_variants(id),
